fixed MSVC compile warnings

+++ kernels/subdiv/feature_adaptive_eval_grid.h
@@ -63,8 +63,8 @@ namespace embree
         Px(Px), Py(Py), Pz(Pz), U(U), V(V), Nx(Nx), Ny(Ny), Nz(Nz), dwidth(dwidth), /*dheight(dheight),*/ count(0)
       {
         assert(swidth < (2<<20) && sheight < (2<<20));
-        const BBox2f srange(Vec2f(0.0f,0.0f),Vec2f(swidth-1,sheight-1));
-        const BBox2f erange(Vec2f(x0,y0),Vec2f(x1,y1));
+        const BBox2f srange(Vec2f(0.0f,0.0f),Vec2f(float(swidth-1),float(sheight-1)));
+        const BBox2f erange(Vec2f((float)x0,(float)y0),Vec2f((float)x1,(float)y1));
         
         /* convert into standard quad patch if possible */
         if (likely(patch.isQuadPatch()))
